Extract and Create Duke Back End Class
